+++ mrv2/docs/HISTORY.md
@@ -2,6 +2,7 @@ v1.1.3
 ======
 
 - pyFLTK build improvements (removal of warnings).
+- Made Windows .zip file smaller by removing unit tests.
 
 
 v1.1.2